--- crates/login/.creds.example
@@ -1,10 +1,0 @@
-first = "firstname"
-last = "lastname" 
-passwd = "$1$<md5 hash of your password>"
-id0 = "<get this from the xmlrpc request>"
-platform = "<lnx, win, or mac>"
-mac = "<your mac address>"
-platform_version = "<whatever you want>"
-start = "last"
-channel = "testclient"
-version = "testclient - test"
