--- crates/login/Cargo.toml
@@ -1,39 +1,0 @@
-[package]
-name = "metaverse_login"
-version = "0.1.2"
-authors = ["Skyler Clark<skylerjaneclark@gmail.com>"]
-description = "login utility for accessing the metaverse"
-edition = "2021"
-license = "AGPL-3.0-or-later"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-xmlrpc = "0.15.1"
-lazy_static = "1.5.0"
-metaverse_instantiator = "0.0.4"
-metaverse_messages = {path = "../messages/"} 
-config = "0.15"
-sys-info = "0.9.1"
-mac_address = "1.1.8"
-md-5 = "0.10.6"
-zip-extract = "0.2.1"
-reqwest = { version = "0.12", features = ["blocking"]}
-rust-ini = "0.21.1"
-log = "0.4"
-env_logger = "0.11"
-actix = "0.13.5"
-actix-rt = "2.10"
-hyper = { version = "1.6" }  
-regex = "1"
-tokio = { version = "1", features = ["full"] }
-[dependencies.uuid]
-version = "1.13.1"
-features = [
-    "v4",                # Lets you generate random UUIDs
-    "fast-rng",          # Use a faster (but still sufficiently random) RNG
-    "macro-diagnostics", # Enable better diagnostics for compile-time UUIDs
-    "serde",
-]
