--- crates/login/README.md
@@ -1,12 +1,0 @@
-Rust implementation of OpenSimulator's [login protocol](http://opensimulator.org/wiki/SimulatorLoginProtocol)
-
-# Getting Started 
-## Using Library 
-### [docs](https://docs.rs/metaverse_login/0.0.0/metaverse_login/) 
-### [crate](https://crates.io/crates/metaverse_login)
-
-add 
-```metaverse_login = "0.1.1"
-
-```
-to your cargo.toml
