--- crates/login/conf.toml
@@ -1,1 +1,0 @@
-channel="unnamed-mmo"
