--- crates/login/sim_config.toml
@@ -1,7 +1,0 @@
-
-#this is the default config for an opensimulator server for testing
-url = "http://opensimulator.org/dist/OpenSim-LastAutoBuild.zip"
-name = "opensim"
-archive = "sim.zip"
-path = "sims"
-executable = "OpenSim.exe"
