--- crates/login/src/models.rs
@@ -1,3 +1,0 @@
-pub mod errors;
-pub mod login_response;
-pub mod simulator_login_protocol;
