--- crates/login/tests/test_server/requirements.txt
@@ -1,1 +1,0 @@
-rawxml
