+++ crates/messages/Cargo.toml
@@ -8,16 +8,20 @@ license = "AGPL-3.0-or-later"
 
 [dependencies]
 log = "0.4"
+xmlrpc = "0.15.1"
 rand = "0.9.0"
 regex = "1"
 byteorder = "1.5.0"
+mac_address = "1.1.8"
 hex = "0.4"
+sys-info = "0.9.1"
 byte = "0.2.7"
 bitflags = "2.8.0"
 metaverse_utils = "0.0.0"
 actix = "0.13.5"
 tokio = { version = "1.43.0", features = ["full"] }
 futures = "0.3"
+md-5 = "0.10.6"
 glam = "0.29.2"
 serde = { version = "1.0", features = ["derive"] }
 [dependencies.uuid]
@@ -26,4 +30,5 @@ features = [
     "v4",                # Lets you generate random UUIDs
     "fast-rng",          # Use a faster (but still sufficiently random) RNG
     "macro-diagnostics", # Enable better diagnostics for compile-time UUIDs
+    "serde",
 ]
