+++ crates/messages/src/chat_from_simulator.rs
