+++ crates/messages/src/chat_from_viewer.rs
