+++ crates/messages/src/coarse_location_update.rs
