+++ crates/messages/src/complete_agent_movement.rs
