+++ crates/messages/src/lib.rs
@@ -1,1 +1,13 @@
-pub mod models;
+pub mod agent_update;
+pub mod chat_from_simulator;
+pub mod chat_from_viewer;
+pub mod circuit_code;
+pub mod client_update_data;
+pub mod coarse_location_update;
+pub mod complete_agent_movement;
+pub mod disable_simulator;
+pub mod header;
+pub mod packet;
+pub mod packet_ack;
+pub mod packet_types;
+pub mod login;
