+++ crates/messages/src/login/login.rs
@@ -1,18 +1,38 @@
-use crate::models::errors::{create_login_error_from_message, LoginError};
-use crate::models::login_response::LoginResponse;
-use crate::models::simulator_login_protocol::{SimulatorLoginOptions, SimulatorLoginProtocol};
-use md5;
-use metaverse_messages::models::login::Login;
+use crate::client_update_data::ClientUpdateData;
+use crate::packet::MessageType;
+use crate::login::errors::{create_login_error_from_message, LoginError, Reason};
+use crate::login::login_response::LoginResponse;
+use crate::login::simulator_login_protocol::{SimulatorLoginOptions, SimulatorLoginProtocol};
 use std::env;
 use std::error::Error;
 
 use mac_address::get_mac_address;
 use md5::{Digest, Md5};
 use std::fs::File;
-use std::io::Read;
 
 extern crate sys_info;
+use crate::header::{Header, PacketFrequency};
+use crate::packet::{Packet, PacketData};
+use futures::future::BoxFuture;
+use std::any::Any;
+use std::collections::HashMap;
+use std::io::{self, BufRead, Read};
+use std::sync::Arc;
+use std::sync::Mutex;
+use tokio::sync::oneshot::Sender;
 
+
+#[derive(Debug, Clone)]
+pub struct Login {
+    pub first: String,
+    pub last: String,
+    pub passwd: String,
+    pub start: String,
+    pub channel: String,
+    pub agree_to_tos: bool,
+    pub read_critical: bool,
+    pub url: String,
+}
 ///Logs in using a SimulatorLoginProtocol object and the url string.
 /// returns a LoginResult, or an error.
 /// If the login response xml can successfully be converted into a LoginResponse struct, do that and
@@ -53,7 +73,7 @@ pub fn login(login_data: SimulatorLoginProtocol, url: String) -> Result<LoginRes
         Ok(request) => request,
         Err(e) => {
             return Err(LoginError::new(
-                crate::models::errors::Reason::Connection,
+                Reason::Connection,
                 &e.to_string(),
             ))
         }
@@ -141,4 +161,98 @@ fn hash_viewer_digest() -> Result<String, Box<dyn Error>> {
     let hash = hasher.finalize();
 
     Ok(format!("{:x}", hash))
+}
+
+impl PacketData for Login {
+    fn from_bytes(bytes: &[u8]) -> io::Result<Self> {
+        let mut cursor = std::io::Cursor::new(bytes);
+
+        let read_string = |cursor: &mut std::io::Cursor<&[u8]>| -> io::Result<String> {
+            let mut buffer = Vec::new();
+            cursor.read_until(0, &mut buffer)?;
+            buffer.pop(); // Remove null terminator
+            Ok(String::from_utf8(buffer)
+                .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?)
+        };
+
+        let first = read_string(&mut cursor)?;
+        let last = read_string(&mut cursor)?;
+        let passwd = read_string(&mut cursor)?;
+        let start = read_string(&mut cursor)?;
+        let channel = read_string(&mut cursor)?;
+
+        let mut bool_buffer = [0u8; 2];
+        cursor.read_exact(&mut bool_buffer)?;
+        let agree_to_tos = bool_buffer[0] != 0;
+        let read_critical = bool_buffer[1] != 0;
+
+        let url = read_string(&mut cursor)?;
+
+        Ok(Login {
+            first,
+            last,
+            passwd,
+            start,
+            channel,
+            agree_to_tos,
+            read_critical,
+            url,
+        })
+    }
+
+    fn to_bytes(&self) -> Vec<u8> {
+        let mut bytes = Vec::new();
+        bytes.extend(self.first.as_bytes());
+        bytes.push(0);
+        bytes.extend(self.last.as_bytes());
+        bytes.push(0);
+        bytes.extend(self.passwd.as_bytes());
+        bytes.push(0);
+        bytes.extend(self.start.as_bytes());
+        bytes.push(0);
+        bytes.extend(self.channel.as_bytes());
+        bytes.push(0);
+        bytes.push(self.agree_to_tos as u8);
+        bytes.push(self.read_critical as u8);
+        bytes.extend(self.url.as_bytes());
+        bytes.push(0);
+        bytes
+    }
+
+    fn on_receive(
+        &self,
+        _: Arc<Mutex<HashMap<u32, Sender<()>>>>,
+        _: Arc<Mutex<Vec<ClientUpdateData>>>,
+    ) -> BoxFuture<'static, ()> {
+        Box::pin(async move {
+            println!("Login packet received");
+        })
+    }
+
+    fn message_type(&self) -> MessageType {
+        MessageType::Login
+    }
+
+    fn as_any(&self) -> &dyn Any {
+        self
+    }
+}
+
+impl Packet {
+    pub fn new_login_packet(login_packet: Login) -> Self {
+        Packet {
+            header: Header {
+                id: 66,
+                frequency: PacketFrequency::Fixed,
+                reliable: true,
+                sequence_number: 0,
+                appended_acks: false,
+                zerocoded: false,
+                resent: false,
+                ack_list: None,
+                size: None,
+            },
+            body: Arc::new(login_packet),
+        }
+    }
 }
