--- crates/messages/src/models/mod.rs
@@ -1,13 +1,0 @@
-pub mod agent_update;
-pub mod chat_from_simulator;
-pub mod chat_from_viewer;
-pub mod circuit_code;
-pub mod client_update_data;
-pub mod coarse_location_update;
-pub mod complete_agent_movement;
-pub mod disable_simulator;
-pub mod header;
-pub mod login;
-pub mod packet;
-pub mod packet_ack;
-pub mod packet_types;
