+++ crates/session/Cargo.toml
@@ -11,7 +11,6 @@ license = "AGPL-3.0-or-later"
 [dependencies]
 config = "0.15"
 metaverse_messages = {path = "../messages/"}
-metaverse_login = {path = "../login/"}
 metaverse_instantiator = "0.0.4"
 regex = "1"
 xmlrpc = "0.15.1"
@@ -19,6 +18,7 @@ actix-web = "4"
 tokio = { version = "1.43.0", features = ["full"] }
 tokio-test = "0.4.4"
 env_logger = "0.11"
+tempfile = "3.17.1"
 log = "0.4"
 actix = "0.13.5"
 hex = "0.4"
@@ -27,6 +27,7 @@ futures = "0.3.31"
 thiserror = "2.0.11"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+crossbeam-channel = "0.5.14"
 
 actix-rt = "2.10"
 [dependencies.uuid]
