+++ crates/ui/Cargo.toml
@@ -7,7 +7,6 @@ edition = "2021"
 bevy = "0.15"
 metaverse_session = {path = "../session/"}
 metaverse_messages = {path = "../messages/"}
-metaverse_login = {path = "../login/"}
 bevy-tokio-tasks = {path = "../../../bevy-actix-tasks/"}
 tokio = { version = "1", features = ["full"] }
 actix-rt = "2.10.0"
